Update script.js
Selected the index file and the paragraph with the balloon, made a function called size, and called new function called set size with a value of newSize. I made size equal new size. I then selected the paragraph again and the css for the paragraph in order to increase the size of the image's pixels. I set the size to 20px