Update script.js
made a function called arrow with an event listener. I created a a if/else/else if statement for the balloon. if you press the up arrow the balloon goes up more and more, unless it is less than 70 pixels. if you press the down arrow it will slowly shrink until it is at its setSize again.